pull 'VirtualQueryEx' into 'STROOP.Win32' project and document intent
