+++ STROOP.Core/GameMemoryAccess/DolphinProcessIO.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.Runtime.InteropServices;
 using Windows.Win32.Foundation;
 using Windows.Win32.System.ProcessStatus;
-using static STROOP.Core.Kernal32NativeMethods;
 
 namespace STROOP.Core.GameMemoryAccess;
 
@@ -16,13 +15,13 @@ public class DolphinProcessIO : WindowsProcessRamIO
 
     protected override void CalculateOffset()
     {
-        MemoryBasicInformation info;
-        IntPtr infoSize = (IntPtr)Marshal.SizeOf(typeof(MemoryBasicInformation));
+        VirtualQueryEx.MemoryBasicInformation info;
+        IntPtr infoSize = (IntPtr)Marshal.SizeOf(typeof(VirtualQueryEx.MemoryBasicInformation));
 
         _baseOffset = (UIntPtr)0;
         bool mem1Found = false;
         for (IntPtr p = new IntPtr();
-             VirtualQueryEx(_processHandle, p, out info, infoSize) == infoSize;
+             VirtualQueryEx.Invoke(_processHandle, p, out info, infoSize) == infoSize;
              p = (IntPtr)(p.ToInt64() + info.RegionSize.ToInt64()))
         {
             if (mem1Found)
@@ -39,7 +38,7 @@ public class DolphinProcessIO : WindowsProcessRamIO
                 continue;
             }
 
-            if (info.RegionSize == (IntPtr)0x2000000 && info.Type == MemoryType.MEM_MAPPED)
+            if (info.RegionSize == (IntPtr)0x2000000 && info.Type == VirtualQueryEx.MemoryType.MEM_MAPPED)
             {
                 // Here, it's likely the right page, but it can happen that multiple pages with these criteria
                 // exists and have nothing to do with the emulated memory. Only the right page has valid
